update stock market

+++ Controllers/StockController.cs
@@ -1,5 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Authorization;
+using  dotnet_api_learning.Helpers;
 using dotnet_api_learning.Data;
 using dotnet_api_learning.Mappers;
 using dotnet_api_learning.Dtos.Stock;
@@ -29,13 +31,25 @@ namespace dotnet_api_learning.Controllers
         // }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        [Authorize]
+        public async Task<IActionResult> GetAll([FromQuery] QueryObject query)
         {
             // var stocks = await _context.Stocks.ToListAsync();
-            var stocks = await _stockRepo.GetAllAsync();
-            var stockDto = stocks.Select(s => s.ToStockDto());
+            var stocks = await _stockRepo.GetAllAsync(query);
+            // var stockDto = stocks.Data.Select(s => s.ToStockDto());
 
-            return Ok(stockDto);
+            // var response = new PagedResponse<StockDto>(
+            //     stockDtos, 
+            //     pagedStocks.TotalRecords, 
+            //     pagedStocks.PageNumber, 
+            //     pagedStocks.PageSize
+            // );
+
+            // return Ok(response);
+            // Cleanly map Stock -> StockDto using the helper method
+            var response = stocks.Map(s => s.ToStockDto());
+
+            return Ok(response);
         }
 
 
@@ -53,9 +67,11 @@ namespace dotnet_api_learning.Controllers
         //     return Ok(stock.ToStockDto());
         // }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var stock = await _stockRepo.GetByIdAsync(id);
 
             if(stock == null)
@@ -80,6 +96,8 @@ namespace dotnet_api_learning.Controllers
         [HttpPost]
         public async Task<IActionResult> CreateStock([FromBody] CreateStockRequestDto stockDto)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var stockModel = stockDto.ToStockFromCreateDto();
             // await _context.Stocks.AddAsync(stockModel);
             // await _context.SaveChangesAsync();
@@ -112,9 +130,11 @@ namespace dotnet_api_learning.Controllers
 
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> UpdateStock([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             // var stockModel = await _context.Stocks.FirstOrDefaultAsync(s => s.Id == id);
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
@@ -154,10 +174,14 @@ namespace dotnet_api_learning.Controllers
         // }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> DeleteStock([FromRoute] int id)
         {
-            // var stockModel = await _context.Stocks.FirstOrDefaultAsync(s => s.Id == id);
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
+            try
+            {
+                            // var stockModel = await _context.Stocks.FirstOrDefaultAsync(s => s.Id == id);
             var stockModel = await _stockRepo.DeleteAsync(id);
 
             if(stockModel == null)
@@ -167,7 +191,14 @@ namespace dotnet_api_learning.Controllers
 
             // _context.Stocks.Remove(stockModel);
             // await _context.SaveChangesAsync();
-            return NoContent();
+            return NoContent();                
+            }
+            catch (DbUpdateException)
+            {
+                
+                return BadRequest("Cannot delete this stock because it has associated comments.");
+            }
+
         }
